Split commands and helps into their own files
